+++ database/migrations/2021_09_19_000001_add_challenge_disable.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
+
+class AddChallengeDisable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('challenges', function ($table) {
+            $table->boolean('disabled')->default(0)->after('pinned');
+        });
+
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('challenges', function ($table) {
+            $table->dropColumn('disabled');
+        });
+
+    }
+}